Replace discover page with landing page
